Implemented code modularity for history feature
