+++ index.html
@@ -8,8 +8,9 @@
   <link rel="icon" type="image/x-icon" href="assets/calculator.png" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
   <link rel="stylesheet" href="style.css" type="text/css" />
-  <script defer src="script.js"></script>
-  <script defer src="regex.js"></script>
+  <script defer type="module" src="script.js"></script>
+  <script defer type="module" src="util/history.js"></script>
+  <!-- <script defer src="regex.js"></script> -->
 </head>
 
 <body>
